MovementEngine: Fixed unit tests
Previously, the ME unit tests manually added players to the PlayerList.
This caused all sorts of problems. Instead, hardcoded in some of the
default card data and use GameState to initialize PlayerList with suspects
from the data file. If there is time, this should be cleaned up further,
but for now it' working.