SuggestionEngine: Use Location for verifying Suggestion validity
Previously, we used location keys provided by Player's get_current_location().
We were unable to validate Suggestions correctly because the Suggestion's
room Card only has the card_id field which represents a room's name, not it's
coordinate key.

Update the various subsystem to pass a Location back from the Board to the
MovementEngine, and finally to the GameController.